refactor(events): keep module helpers below size cap
