+++ crates/perry-ext-events/src/lib.rs
@@ -32,7 +32,14 @@ mod error_monitor;
 use error_monitor::dispatch_error_monitor;
 mod max_listeners;
 mod messages;
+mod module_helpers;
 mod target_helpers;
+
+use module_helpers::{call_net_socket_method, js_events_native_dispatch};
+pub use module_helpers::{
+    js_events_get_event_listeners, js_events_get_max_listeners, js_events_listener_count,
+    js_events_set_max_listeners,
+};
 
 #[cfg(test)]
 mod test_async_shims;
@@ -154,12 +161,6 @@ extern "C" {
     fn js_implicit_this_set(value: f64) -> f64;
     fn js_jsvalue_to_string(value: f64) -> *mut StringHeader;
     fn js_native_call_value(func_value: f64, args_ptr: *const f64, args_len: usize) -> f64;
-    fn js_native_call_method_str_key(
-        object: f64,
-        name_handle: i64,
-        args_ptr: *const f64,
-        args_len: usize,
-    ) -> f64;
     fn js_value_is_promise(value: f64) -> i32;
     fn js_register_event_emitter_handle_probe(f: unsafe extern "C" fn(i64) -> bool);
     fn js_register_event_emitter_get_domain(f: unsafe extern "C" fn(i64) -> i64);
@@ -242,16 +243,6 @@ impl EventsRootVisitor {
 #[inline]
 fn nanbox_pointer_bits(ptr: i64) -> f64 {
     f64::from_bits(POINTER_TAG | ((ptr as u64) & POINTER_MASK))
-}
-
-unsafe fn call_net_socket_method(handle: Handle, name: &str, args: &[f64]) -> f64 {
-    let name_ptr = js_string_from_bytes(name.as_ptr(), name.len() as u32);
-    js_native_call_method_str_key(
-        nanbox_pointer_bits(handle),
-        name_ptr as i64,
-        args.as_ptr(),
-        args.len(),
-    )
 }
 
 /// One registered listener: the original raw closure pointer plus an optional
@@ -1883,238 +1874,6 @@ pub unsafe extern "C" fn js_events_add_abort_listener(signal: f64, listener: f64
     let dispose_sym_val = js_symbol_for(dispose_key_val);
     js_object_set_symbol_property(disposable_val, dispose_sym_val, dispose_val);
     disposable as i64
-}
-
-/// `events.getEventListeners(emitter, eventName)` — alias for
-/// `emitter.listeners(eventName)`.
-///
-/// # Safety
-///
-/// `event_name_ptr` must be null or a Perry-runtime `StringHeader`.
-#[no_mangle]
-pub unsafe extern "C" fn js_events_get_event_listeners(
-    target_value: f64,
-    event_name_ptr: *const StringHeader,
-) -> *mut ArrayHeader {
-    // AbortSignal is an EventTarget in Node, but Perry represents it as its
-    // own native object that `event_helper_target` doesn't recognize. A
-    // signal only ever tracks "abort" listeners.
-    let signal_ptr = js_abort_signal_resolve_ptr(target_value);
-    if !signal_ptr.is_null() {
-        if string_from_header(event_name_ptr).as_deref() == Some("abort") {
-            return js_abort_signal_listeners_copy(signal_ptr);
-        }
-        return js_array_alloc(0);
-    }
-    match event_helper_target(target_value).unwrap_or_else(|| {
-        throw_invalid_arg_type(&invalid_instance_arg_message(
-            "emitter",
-            "EventEmitter or EventTarget",
-            target_value,
-        ))
-    }) {
-        EventHelperTarget::EventEmitter(handle) => {
-            js_event_emitter_listeners(handle, event_name_ptr as i64)
-        }
-        EventHelperTarget::EventTarget(target) => {
-            js_event_target_get_event_listeners(target, event_name_ptr)
-        }
-        EventHelperTarget::NetSocket(handle) | EventHelperTarget::NativeHandle(handle) => {
-            let event = f64::from_bits(nanbox_string_bits(event_name_ptr as *mut StringHeader));
-            let value = call_net_socket_method(handle, "listeners", &[event]);
-            (value.to_bits() & POINTER_MASK) as *mut ArrayHeader
-        }
-        EventHelperTarget::Stream(handle) => {
-            stream_listeners_for_heap_object(handle, event_name_ptr)
-                .unwrap_or_else(|| js_array_alloc(0))
-        }
-    }
-}
-
-/// `events.listenerCount(emitter, eventName)` — alias.
-///
-/// # Safety
-///
-/// `event_name_ptr` must be null or a Perry-runtime `StringHeader`.
-#[no_mangle]
-pub unsafe extern "C" fn js_events_listener_count(
-    target_value: f64,
-    event_name_ptr: *const StringHeader,
-) -> f64 {
-    // AbortSignal: see `js_events_get_event_listeners`.
-    let signal_ptr = js_abort_signal_resolve_ptr(target_value);
-    if !signal_ptr.is_null() {
-        if string_from_header(event_name_ptr).as_deref() == Some("abort") {
-            return js_abort_signal_listener_count(signal_ptr);
-        }
-        return 0.0;
-    }
-    match event_helper_target(target_value).unwrap_or_else(|| {
-        throw_invalid_arg_type(&invalid_instance_arg_message(
-            "emitter",
-            "EventEmitter or EventTarget",
-            target_value,
-        ))
-    }) {
-        EventHelperTarget::EventEmitter(handle) => js_event_emitter_listener_count(
-            handle,
-            event_name_ptr as i64,
-            TAG_UNDEFINED_F64_BITS as i64,
-        ),
-        EventHelperTarget::EventTarget(target) => event_target_array_len(target, event_name_ptr),
-        EventHelperTarget::NetSocket(handle) | EventHelperTarget::NativeHandle(handle) => {
-            let event = f64::from_bits(nanbox_string_bits(event_name_ptr as *mut StringHeader));
-            call_net_socket_method(handle, "listenerCount", &[event])
-        }
-        EventHelperTarget::Stream(handle) => {
-            stream_array_len(handle, event_name_ptr).unwrap_or(0.0)
-        }
-    }
-}
-
-/// `events.getMaxListeners(emitter)` — alias.
-#[no_mangle]
-pub unsafe extern "C" fn js_events_get_max_listeners(target_value: f64) -> f64 {
-    // AbortSignal: Node's default EventTarget listener cap. Perry stores no
-    // per-signal override (`setMaxListeners` below is an accepted no-op), so
-    // the default is always reported.
-    if !js_abort_signal_resolve_ptr(target_value).is_null() {
-        return 10.0;
-    }
-    match event_helper_target(target_value).unwrap_or_else(|| {
-        throw_invalid_arg_type(&invalid_instance_arg_message(
-            "emitter",
-            "EventEmitter or EventTarget",
-            target_value,
-        ))
-    }) {
-        EventHelperTarget::EventEmitter(handle) => js_event_emitter_get_max_listeners(handle),
-        EventHelperTarget::EventTarget(target) => js_event_target_get_max_listeners(target),
-        EventHelperTarget::NetSocket(_) | EventHelperTarget::NativeHandle(_) => 10.0,
-        EventHelperTarget::Stream(handle) => js_node_stream_method_get_max_listeners(handle),
-    }
-}
-
-/// `events.setMaxListeners(n, ...emitters)` — Perry FFI takes a single
-/// array of target handles from the codegen varargs lowering.
-#[no_mangle]
-pub unsafe extern "C" fn js_events_set_max_listeners(
-    n: f64,
-    handles_ptr: *const ArrayHeader,
-) -> f64 {
-    let n = validate_max_listeners(n);
-    if !handles_ptr.is_null() {
-        let len = (*handles_ptr).length;
-        for i in 0..len {
-            let value = f64::from_bits(js_array_get(handles_ptr, i).bits());
-            // AbortSignal is an EventTarget in Node — SDKs routinely call
-            // `events.setMaxListeners(n, controller.signal)` to raise the
-            // MaxListenersExceededWarning threshold on a shared signal. Perry
-            // represents signals as their own native object that
-            // `event_helper_target` doesn't recognize, so this threw
-            // ERR_INVALID_ARG_TYPE and rejected the caller's whole request
-            // path. Accept the signal; the warning threshold is the call's
-            // only Node-observable effect and Perry never emits that warning
-            // for signals, so accepting is a faithful no-op.
-            if !js_abort_signal_resolve_ptr(value).is_null() {
-                continue;
-            }
-            match event_helper_target(value).unwrap_or_else(|| {
-                throw_invalid_arg_type(&invalid_instance_arg_message(
-                    "eventTargets",
-                    "EventEmitter or EventTarget",
-                    value,
-                ))
-            }) {
-                EventHelperTarget::EventEmitter(handle) => {
-                    if let Some(emitter) = get_event_emitter_mut(handle) {
-                        emitter.max_listeners = n;
-                    }
-                }
-                EventHelperTarget::EventTarget(target) => {
-                    let _ = js_event_target_set_max_listeners(target, n);
-                }
-                EventHelperTarget::NetSocket(_) | EventHelperTarget::NativeHandle(_) => {}
-                EventHelperTarget::Stream(handle) => {
-                    let _ = js_node_stream_method_set_max_listeners(handle, n);
-                }
-            }
-        }
-    }
-    f64::from_bits(0x7FFC_0000_0000_0001)
-}
-
-unsafe fn event_name_header(value: f64) -> *const StringHeader {
-    js_jsvalue_to_string(value)
-}
-
-/// Indirect/captured `events.*` helper dispatcher for the external events
-/// implementation. Without this registration, CommonJS destructuring such as
-/// `const { getEventListeners } = require('events')` resolves the callable but
-/// returns `undefined` when invoked.
-#[no_mangle]
-pub unsafe extern "C" fn js_events_native_dispatch(
-    method: *const u8,
-    method_len: usize,
-    args: *const f64,
-    args_len: usize,
-) -> f64 {
-    let undefined = f64::from_bits(TAG_UNDEFINED_F64_BITS);
-    if method.is_null() || method_len == 0 {
-        return undefined;
-    }
-    let name = std::str::from_utf8(std::slice::from_raw_parts(method, method_len)).unwrap_or("");
-    let arg = |index: usize| {
-        if !args.is_null() && index < args_len {
-            *args.add(index)
-        } else {
-            undefined
-        }
-    };
-    match name {
-        "listenerCount" => js_events_listener_count(arg(0), event_name_header(arg(1))),
-        "getMaxListeners" => js_events_get_max_listeners(arg(0)),
-        "getEventListeners" => {
-            let array = js_events_get_event_listeners(arg(0), event_name_header(arg(1)));
-            if array.is_null() {
-                undefined
-            } else {
-                nanbox_pointer_bits(array as i64)
-            }
-        }
-        "once" => {
-            let promise = js_events_once(arg(0), event_name_header(arg(1)), arg(2));
-            if promise.is_null() {
-                undefined
-            } else {
-                nanbox_pointer_bits(promise as i64)
-            }
-        }
-        "on" => {
-            let iterator = js_events_on(arg(0), event_name_header(arg(1)), arg(2));
-            if iterator.is_null() {
-                undefined
-            } else {
-                nanbox_pointer_bits(iterator as i64)
-            }
-        }
-        "addAbortListener" => {
-            let disposable = js_events_add_abort_listener(arg(0), arg(1));
-            if disposable == 0 {
-                undefined
-            } else {
-                nanbox_pointer_bits(disposable)
-            }
-        }
-        "setMaxListeners" => {
-            let mut targets = js_array_alloc(args_len.saturating_sub(1) as u32);
-            for index in 1..args_len {
-                targets = js_array_push_f64(targets, arg(index));
-            }
-            js_events_set_max_listeners(arg(0), targets)
-        }
-        _ => undefined,
-    }
 }
 
 #[cfg(test)]
